requirements.txt
torch
torchvision
numpy
matplotlib
scikit-learn
seaborn
pillow
opencv-python